+++ playwright/tests/click.spec.ts
@@ -9,3 +9,13 @@ When(
     return await this.page.locator(element).click();
   }
 );
+
+When(
+  /^I select "([^"]*)" from the category dropdown$/,
+  async function (this: CustomWorld, elementKey: keyof typeof elements) {
+    const searchCategory = this.page.locator(
+      elements["search category dropdown"]
+    );
+    return await searchCategory.selectOption(elementKey);
+  }
+);