+++ src/components/ReaderLayout/ReaderLayout.tsx
@@ -1,11 +1,10 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { Breadcrumbs, Icon } from "@nypl/design-system-react-components";
 import { defaultBreadcrumbs } from "~/src/constants/labels";
 import { ApiLink, LinkResult } from "~/src/types/LinkQuery";
 import IFrameReader from "../IFrameReader/IFrameReader";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import Layout from "../Layout/Layout";
-import * as gtag from "../../lib/Analytics";
 import { formatUrl, truncateStringOnWhitespace } from "~/src/util/Util";
 import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
 import dynamic from "next/dynamic";
@@ -24,6 +23,9 @@ import readiumDefault from "!file-loader!extract-loader!css-loader!@nypl/web-rea
 import readiumAfter from "!file-loader!extract-loader!css-loader!@nypl/web-reader/dist/injectable-html-styles/ReadiumCSS-after.css";
 
 import Link from "../Link/Link";
+import { addTocToManifest } from "@nypl/web-reader";
+import Loading from "../Loading/Loading";
+import { trackCtaClick } from "~/src/lib/Analytics";
 
 const origin =
   typeof window !== "undefined" && window.location?.origin
@@ -57,16 +59,81 @@ const ReaderLayout: React.FC<{
   backUrl: string;
 }> = (props) => {
   const link: ApiLink = props.linkResult.data;
-  const proxyUrl = props.proxyUrl;
   const url = formatUrl(link.url);
+  const proxyUrl = props.proxyUrl;
   const edition = link.work.editions[0];
+  const [manifestUrl, setManifestUrl] = useState(url);
+  const [isLoading, setIsLoading] = useState(true);
 
   const isEmbed = MediaTypes.embed.includes(link.media_type);
   const isRead = MediaTypes.read.includes(link.media_type);
 
+  const pdfWorkerSrc = `${origin}/pdf-worker/pdf.worker.min.js`;
+
+  /**
+   * This is a function we will use to get the resource through a given proxy url.
+   * It will eventually be passed to the web reader instead of passing a proxy url directly.
+   */
+  const getProxiedResource = (proxyUrl?: string) => async (href: string) => {
+    // Generate the resource URL using the proxy
+    const url: string = proxyUrl
+      ? `${proxyUrl}${encodeURIComponent(href)}`
+      : href;
+    const response = await fetch(url, { mode: "cors" });
+    const array = new Uint8Array(await response.arrayBuffer());
+
+    if (!response.ok) {
+      throw new Error("Response not Ok for URL: " + url);
+    }
+    return array;
+  };
+
   useEffect(() => {
-    gtag.drbEvents("Read", `${link.work.title}`);
+    trackCtaClick({
+      cta_section: `${link.work.title}`,
+      cta_text: "Read Online",
+      destination_url: `${link.url}`,
+    });
   }, [link]);
+
+  useEffect(() => {
+    if (isRead) {
+      /**
+       * - Fetches manifest
+       * - Adds the TOC to the manifest
+       * - Generates a syncthetic url for the manifest to be passed to
+       * web reader.
+       * - Returns the synthetic url
+       */
+      const fetchAndModifyManifest = async (url) => {
+        setIsLoading(true);
+        const response = await fetch(url);
+        const manifest = await response.json();
+        if (
+          manifest &&
+          manifest.readingOrder &&
+          manifest.readingOrder.length === 1
+        ) {
+          const modifiedManifest = await addTocToManifest(
+            manifest,
+            getProxiedResource(proxyUrl),
+            pdfWorkerSrc
+          );
+          const syntheticUrl = URL.createObjectURL(
+            new Blob([JSON.stringify(modifiedManifest)])
+          );
+          setManifestUrl(syntheticUrl);
+        }
+        setIsLoading(false);
+      };
+
+      fetchAndModifyManifest(url);
+
+      // hides header and footer components when web reader is displayed
+      document.getElementById("nypl-header").style.display = "none";
+      document.getElementById("nypl-footer").style.display = "none";
+    }
+  }, [isRead, pdfWorkerSrc, proxyUrl, url]);
 
   const BackButton = () => {
     return (
@@ -104,11 +171,12 @@ const ReaderLayout: React.FC<{
           <IFrameReader url={link.url} />
         </Layout>
       )}
-      {isRead && (
+      {isRead && isLoading && <Loading />}
+      {isRead && !isLoading && (
         <WebReader
-          webpubManifestUrl={url}
+          webpubManifestUrl={manifestUrl}
           proxyUrl={proxyUrl}
-          pdfWorkerSrc={`${origin}/pdf-worker/pdf.worker.min.js`}
+          pdfWorkerSrc={pdfWorkerSrc}
           headerLeft={<BackButton />}
           injectablesFixed={injectables}
         />